🏗️  Refactor: let app in an separeted file
